Changed floating point precision to f32 as there is no need to use f64 here

+++ data-generator-cli/src/main.rs
@@ -57,12 +57,12 @@ fn generate_lines(count: u32) -> Vec<String> {
     );
 
     for i in 0..count {
-        let base_price = rng.gen::<f64>() * 1000.0f64;
+        let base_price = rng.gen::<f32>() * 1000.0f32;
 
         let index = i.to_string();
         let name = item_names[rng.gen_range(0..item_names.len())].to_owned();
         let min_price = format!("{:.2}", base_price);
-        let max_price = format!("{:.2}", ((rng.gen::<f64>() * (base_price * 0.5f64)) + base_price));
+        let max_price = format!("{:.2}", ((rng.gen::<f32>() * (base_price * 0.5f32)) + base_price));
 
         lines.push(vec!(index, name, min_price, max_price).join(", "));
     }
